+++ lib/stac.rb
@@ -1,7 +1,12 @@
 # frozen_string_literal: true
 
-require_relative 'stac/simple_http_client'
 require_relative 'stac/object_resolver'
+require_relative 'stac/simple_http_client'
+require_relative 'stac/stac_object'
+require_relative 'stac/extensions/electro_optical'
+require_relative 'stac/extensions/projection'
+require_relative 'stac/extensions/scientific_citation'
+require_relative 'stac/extensions/view_geometry'
 require_relative 'stac/version'
 
 # Gem namespace.
@@ -27,4 +32,9 @@ module STAC
   end
 
   self.default_http_client = SimpleHTTPClient.new
+
+  STACObject.add_extendable(Extensions::ElectroOptical)
+  STACObject.add_extendable(Extensions::Projection)
+  STACObject.add_extendable(Extensions::ScientificCitation)
+  STACObject.add_extendable(Extensions::ViewGeometry)
 end