+++ lib/stac/catalog.rb
@@ -12,11 +12,11 @@ module STAC
 
     attr_accessor :id, :description, :title
 
-    def initialize(id:, description:, links:, title: nil, stac_extensions: nil, **extra)
-      super(links: links, stac_extensions: stac_extensions, **extra)
+    def initialize(id:, description:, links:, title: nil, stac_extensions: [], **extra)
       @id = id
       @description = description
       @title = title
+      super(links: links, stac_extensions: stac_extensions, **extra)
     end
 
     # Serializes self to a Hash.