+++ lib/stac/stac_object.rb
@@ -8,6 +8,18 @@ require_relative 'spec_version'
 module STAC
   # Base class for \STAC objects (i.e. Catalog, Collection and Item).
   class STACObject
+    @@extendables = {} # rubocop:disable Style/ClassVars
+
+    # Returns available extension modules.
+    def self.extendables
+      @@extendables.values.uniq
+    end
+
+    # Adds the given extension module to .extendables.
+    def self.add_extendable(extendable)
+      @@extendables[extendable.identifier] = extendable
+    end
+
     class << self
       attr_accessor :type # :nodoc:
 
@@ -18,28 +30,26 @@ module STAC
         raise TypeError, "type field is not '#{type}': #{hash['type']}" if hash.fetch('type') != type
 
         transformed = hash.transform_keys(&:to_sym).except(:type, :stac_version)
-        transformed[:links] = transformed.fetch(:links).map { |link| Link.from_hash(link) }
         new(**transformed)
-      rescue KeyError => e
-        raise ArgumentError, "required field not found: #{e.key}"
       end
     end
 
-    attr_accessor :stac_extensions, :extra
+    attr_accessor :extra
 
     # HTTP Client to fetch objects from HTTP HREF links.
     attr_accessor :http_client
 
-    attr_reader :links
+    attr_reader :stac_extensions, :links
 
-    def initialize(links:, stac_extensions: nil, **extra)
+    def initialize(links:, stac_extensions: [], **extra)
       @links = []
       links.each do |link|
-        add_link(link) # to set `owner`
+        add_link(**link.transform_keys(&:to_sym)) # to set `owner`
       end
       @stac_extensions = stac_extensions
       @extra = extra.transform_keys(&:to_s)
       @http_client = STAC.default_http_client
+      apply_extensions!
     end
 
     def type
@@ -51,7 +61,7 @@ module STAC
       {
         'type' => type,
         'stac_version' => SPEC_VERSION,
-        'stac_extensions' => stac_extensions,
+        'stac_extensions' => stac_extensions.empty? ? nil : stac_extensions,
         'links' => links.map(&:to_h),
       }.merge(extra).compact
     end
@@ -61,8 +71,30 @@ module STAC
       to_h.to_json(...)
     end
 
+    # Returns extended extension modules.
+    def extended
+      @extended ||= []
+    end
+
+    # Adds the given extension identifier to #stac_extensions.
+    #
+    # When the given argument is extendable, the item extends the module.
+    def add_extension(extension)
+      case extension
+      when Extension
+        stac_extensions << extension.identifier
+        apply_extension!(extension)
+      else
+        stac_extensions << extension
+        if (extension = @@extendables[extension]) && extension.scope.include?(self.class)
+          apply_extension!(extension)
+        end
+      end
+    end
+
     # Adds a link with setting Link#owner as self.
-    def add_link(link)
+    def add_link(rel:, href:, type: nil, title: nil, **extra)
+      link = Link.new(rel: rel, href: href, type: type, title: title, **extra)
       link.owner = self
       links << link
     end
@@ -76,9 +108,8 @@ module STAC
     #
     # When any ref="self" links already exist, removes them.
     def self_href=(absolute_href)
-      self_link = Link.new(rel: 'self', href: absolute_href, type: 'application/json')
       remove_link(rel: 'self')
-      add_link(self_link)
+      add_link(rel: 'self', href: absolute_href, type: 'application/json')
     end
 
     # Returns a link matching the arguments.
@@ -93,6 +124,23 @@ module STAC
     end
 
     private
+
+    def extensions
+      stac_extensions
+        .map { |extension_id| @@extendables[extension_id] }
+        .compact
+        .select { |extension| extension.scope.include?(self.class) }
+    end
+
+    def apply_extensions!
+      extensions.each do |extension|
+        apply_extension!(extension)
+      end
+    end
+
+    def apply_extension!(extension)
+      extended << extension
+    end
 
     def remove_link(rel:)
       links.reject! { |link| link.rel == rel }